+++ internal/overlord/sdkstate/handlers.go
@@ -103,7 +103,7 @@ func (m *SdkManager) doRetrieveSdk(task *state.Task, tomb *tomb.Tomb) error {
 	st := task.State()
 	reporter := &progress.Reporter{
 		Name: task.ID(),
-		Report: func(label string, done, total int) {
+		Report: func(label string, done, total int64) {
 			st.Lock()
 			task.SetProgress(label, done, total)
 			st.Unlock()
